python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flash-arb-bot"
version = "0.1.0"
description = "High-frequency arbitrage trading bot for cryptocurrency futures"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Flash Arbitrage Team"}
]
keywords = ["trading", "arbitrage", "cryptocurrency", "futures"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "redis[hiredis]>=5.0.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "python-binance>=1.0.17",
    "ccxt>=4.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "colorama>=0.4.6",
    "prometheus-client>=0.19.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
flash-arb = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
